refactor: get ipfs endpoint from env, update types
